+++ core/popup/popup.css
@@ -1,4 +1,4 @@
-@import "../../feature-reading-list/reading-list.css";
+@import "../../features/reading-list/reading-list.css";
 @import "../../features/status-bar/status-bar.css";
 
 * {
