+++ core/popup/popup.js
@@ -1,5 +1,5 @@
 import '../../modules/prototype.mjs'
-import * as readingList from '../../feature-reading-list/readingList.mjs'
+import * as readingList from '../../features/reading-list/readingList.mjs'
 import * as statusBar from '../../features/status-bar/statusBar.js'
 
 $(async () => {
