+++ features/reading-list/reading-list.css
@@ -45,7 +45,7 @@
     position: absolute;
     height: 16px;
     width: 16px;
-    background-image: url("../assets/icons/logo-black32x32.png");
+    background-image: url("../../assets/icons/logo-black32x32.png");
     background-size: 16px 16px;
 }
